Extracted a base class from the DOWConditionHandlerTest.
The DayOfWeekConditionHandlerTest provides a lot of functionality that may
be used in the TimeOfDayConditionHandlerTest to test feature #6465. In order to avoid duplicate code, a base class has been extracted.

Also-by: Kai Kreuzer <kai@openhab.org>
